fix: update observation space data types in Crafter class to use float32 for is_first, is_last, and is_terminal

+++ envs/crafter.py
@@ -19,9 +19,9 @@ class Crafter:
             "image": gym.spaces.Box(
                 0, 255, self._env.observation_space.shape, dtype=np.uint8
             ),
-            "is_first": gym.spaces.Box(-np.inf, np.inf, (1,), dtype=np.uint8),
-            "is_last": gym.spaces.Box(-np.inf, np.inf, (1,), dtype=np.uint8),
-            "is_terminal": gym.spaces.Box(-np.inf, np.inf, (1,), dtype=np.uint8),
+            "is_first": gym.spaces.Box(-np.inf, np.inf, (1,), dtype=np.float32),
+            "is_last": gym.spaces.Box(-np.inf, np.inf, (1,), dtype=np.float32),
+            "is_terminal": gym.spaces.Box(-np.inf, np.inf, (1,), dtype=np.float32),
             "log_reward": gym.spaces.Box(-np.inf, np.inf, (1,), dtype=np.float32),
         }
         spaces.update(
